% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_change_in_ds_var}
\alias{add_change_in_ds_var}
\title{Add change in dataset variable}
\usage{
add_change_in_ds_var(
  ds_tb,
  id_var_nm_1L_chr = "fkClientID",
  round_var_nm_1L_chr = "round",
  round_bl_val_1L_chr = "Baseline",
  change_var_nm_1L_chr,
  var_nm_1L_chr,
  arrange_by_id_lgl = T
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one), Default: 'fkClientID'}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one), Default: 'round'}

\item{round_bl_val_1L_chr}{Round baseline value (a character vector of length one), Default: 'Baseline'}

\item{change_var_nm_1L_chr}{Change variable name (a character vector of length one)}

\item{var_nm_1L_chr}{Variable name (a character vector of length one)}

\item{arrange_by_id_lgl}{Arrange by identity (a logical vector), Default: T}
}
\value{
Updated dataset (a tibble)
}
\description{
add_change_in_ds_var() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add change in dataset variable. The function returns Updated dataset (a tibble).
}
\keyword{internal}
