% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_costs_from_gamma_dist}
\alias{add_costs_from_gamma_dist}
\title{Add costs from gamma dist}
\usage{
add_costs_from_gamma_dist(
  ds_tb,
  costs_mean_dbl,
  costs_sd_dbl,
  costs_var_nm_1L_dbl = "costs_dbl"
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{costs_mean_dbl}{Costs mean (a double vector)}

\item{costs_sd_dbl}{Costs sd (a double vector)}

\item{costs_var_nm_1L_dbl}{Costs var name (a double vector of length one), Default: 'costs_dbl'}
}
\value{
Updated dataset (a tibble)
}
\description{
add_costs_from_gamma_dist() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add costs from gamma dist. Function argument ds_tb specifies the object to be updated. The function returns Updated dataset (a tibble).
}
\keyword{internal}
