% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_diffs_by_group_and_tmpt}
\alias{add_diffs_by_group_and_tmpt}
\title{Add differences by group and time point}
\usage{
add_diffs_by_group_and_tmpt(
  ds_tb = trial_ds_tb,
  cmprsn_var_nm_1L_chr = "study_arm_chr",
  cmprsn_group_match_val_chr = c("Intervention"),
  round_var_nm_1L_chr = "round",
  timepoint_match_val_1L_chr = "Follow-up",
  match_idx_var_nm_1L_chr = "match_idx_int",
  var_nms_chr,
  fns_ls,
  abs_mean_diff_dbl,
  diff_sd_dbl,
  multiplier_dbl,
  min_dbl,
  max_dbl,
  integer_lgl
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble), Default: trial_ds_tb}

\item{cmprsn_var_nm_1L_chr}{Comparison variable name (a character vector of length one), Default: 'study_arm_chr'}

\item{cmprsn_group_match_val_chr}{Comparison group match value (a character vector), Default: c("Intervention")}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one), Default: 'round'}

\item{timepoint_match_val_1L_chr}{Timepoint match value (a character vector of length one), Default: 'Follow-up'}

\item{match_idx_var_nm_1L_chr}{Match index variable name (a character vector of length one), Default: 'match_idx_int'}

\item{var_nms_chr}{Variable names (a character vector)}

\item{fns_ls}{Functions (a list)}

\item{abs_mean_diff_dbl}{Absolute mean difference (a double vector)}

\item{diff_sd_dbl}{Difference standard deviation (a double vector)}

\item{multiplier_dbl}{Multiplier (a double vector)}

\item{min_dbl}{Minimum (a double vector)}

\item{max_dbl}{Maximum (a double vector)}

\item{integer_lgl}{Integer (a logical vector)}
}
\value{
Updated dataset (a tibble)
}
\description{
add_diffs_by_group_and_tmpt() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add differences by group and time point. Function argument ds_tb specifies the object to be updated. The function returns Updated dataset (a tibble).
}
\keyword{internal}
