% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_qalys}
\alias{add_qalys}
\title{Add qalys}
\usage{
add_qalys(
  ds_tb,
  cmprsn_var_nm_1L_chr = "study_arm_chr",
  duration_var_nm_1L_chr = "duration_prd",
  id_var_nm_1L_chr = "fkClientID",
  match_idx_var_nm_1L_chr = "match_idx_int",
  qalys_var_nm_1L_chr = "qalys_dbl",
  round_var_nm_1L_chr = "round",
  utl_change_var_nm_1L_chr = "utl_change_dbl",
  utl_var_nm_1L_chr = "utility_dbl",
  reshape_1L_lgl = T
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{cmprsn_var_nm_1L_chr}{Cmprsn var name (a character vector of length one), Default: 'study_arm_chr'}

\item{duration_var_nm_1L_chr}{Duration var name (a character vector of length one), Default: 'duration_prd'}

\item{id_var_nm_1L_chr}{Id var name (a character vector of length one), Default: 'fkClientID'}

\item{match_idx_var_nm_1L_chr}{Match index var name (a character vector of length one), Default: 'match_idx_int'}

\item{qalys_var_nm_1L_chr}{Qalys var name (a character vector of length one), Default: 'qalys_dbl'}

\item{round_var_nm_1L_chr}{Round var name (a character vector of length one), Default: 'round'}

\item{utl_change_var_nm_1L_chr}{Utl change var name (a character vector of length one), Default: 'utl_change_dbl'}

\item{utl_var_nm_1L_chr}{Utl var name (a character vector of length one), Default: 'utility_dbl'}

\item{reshape_1L_lgl}{Reshape (a logical vector of length one), Default: T}
}
\value{
Updated dataset (a tibble)
}
\description{
add_qalys() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add qalys. Function argument ds_tb specifies the object to be updated. The function returns Updated dataset (a tibble).
}
\keyword{internal}
