% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_qalys_to_ds}
\alias{add_qalys_to_ds}
\title{Add Quality Adjusted Life Years to dataset}
\usage{
add_qalys_to_ds(
  ds_tb,
  predn_ds_ls,
  include_predrs_1L_lgl = T,
  reshape_1L_lgl = T
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{predn_ds_ls}{Prediction dataset (a list)}

\item{include_predrs_1L_lgl}{Include predictors (a logical vector of length one), Default: T}

\item{reshape_1L_lgl}{Reshape (a logical vector of length one), Default: T}
}
\value{
Dataset (a tibble)
}
\description{
add_qalys_to_ds() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add quality adjusted life years to dataset. Function argument ds_tb specifies the object to be updated. The function returns Dataset (a tibble).
}
