% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_utl_predn}
\alias{add_utl_predn}
\title{Add utility prediction}
\usage{
add_utl_predn(
  data_tb,
  predn_ds_ls,
  deterministic_1L_lgl = T,
  force_min_max_1L_lgl = T,
  key_1L_chr = NULL,
  make_from_tbl_1L_lgl = T,
  model_mdl = NULL,
  new_data_is_1L_chr = "Simulated",
  server_1L_chr = "dataverse.harvard.edu",
  utl_cls_fn = NULL
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{predn_ds_ls}{Prediction dataset (a list)}

\item{deterministic_1L_lgl}{Deterministic (a logical vector of length one), Default: T}

\item{force_min_max_1L_lgl}{Force minimum maximum (a logical vector of length one), Default: T}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{make_from_tbl_1L_lgl}{Make from table (a logical vector of length one), Default: T}

\item{model_mdl}{Model (a model), Default: NULL}

\item{new_data_is_1L_chr}{New data is (a character vector of length one), Default: 'Simulated'}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}

\item{utl_cls_fn}{Utility class (a function), Default: NULL}
}
\value{
Updated (a tibble)
}
\description{
add_utl_predn() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add utility prediction. The function returns Updated (a tibble).
}
