% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_dictionary_tb.R
\docType{data}
\name{dictionary_tb}
\alias{dictionary_tb}
\title{Data dictionary for study population dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 85 rows and 4 columns.
}
\usage{
dictionary_tb
}
\description{
A data dictionary of the variables used in the source and replication (synthetic) datasets for the First Bounce transfer to utility study
}
\details{
A tibble

\describe{
\item{var_nm_chr}{Var name (a character vector)}
\item{var_cat_chr}{Var cat (a character vector)}
\item{var_desc_chr}{Var description (a character vector)}
\item{var_type_chr}{Var type (a character vector)}
}
}
\keyword{datasets}
