% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_sofas.R
\name{firstbounce_sofas}
\alias{firstbounce_sofas}
\title{First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)}
\usage{
firstbounce_sofas(x = make_pt_firstbounce_sofas())
}
\arguments{
\item{x}{A prototype for the First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS), Default: make_pt_firstbounce_sofas()}
}
\value{
A validated instance of the First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\description{
Create a new valid instance of the First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\details{
First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
