% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_dss_using_predrs}
\alias{get_dss_using_predrs}
\title{Get datasets using predictors}
\usage{
get_dss_using_predrs(
  mdl_predrs_in_ds_chr,
  ttu_dv_dss_tb = NULL,
  ttu_dv_nm_1L_chr = "firstbounce",
  server_1L_chr = "dataverse.harvard.edu",
  key_1L_chr = NULL
)
}
\arguments{
\item{mdl_predrs_in_ds_chr}{Model predictors in dataset (a character vector)}

\item{ttu_dv_dss_tb}{Ttu dataverse datasets (a tibble), Default: NULL}

\item{ttu_dv_nm_1L_chr}{Ttu dataverse name (a character vector of length one), Default: 'firstbounce'}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\value{
Ttu dataverse datasets (a tibble)
}
\description{
get_dss_using_predrs() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get datasets using predictors. Function argument mdl_predrs_in_ds_chr specifies the where to look for the required object. The function returns Ttu dataverse datasets (a tibble).
}
\keyword{internal}
