% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_dv_ds_publication}
\alias{get_dv_ds_publication}
\title{Get dataverse dataset publication}
\usage{
get_dv_ds_publication(
  ds_url_1L_chr,
  server_1L_chr = "dataverse.harvard.edu",
  key_1L_chr = NULL
)
}
\arguments{
\item{ds_url_1L_chr}{Dataset url (a character vector of length one)}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\value{
Digital object identifier url (a character vector of length one)
}
\description{
get_dv_ds_publication() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get dataverse dataset publication. Function argument ds_url_1L_chr specifies the where to look for the required object. The function returns Digital object identifier url (a character vector of length one).
}
\keyword{internal}
