% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_dv_dss_mdl_smrys}
\alias{get_dv_dss_mdl_smrys}
\title{Get dataverse datasets model summarys}
\usage{
get_dv_dss_mdl_smrys(
  ids_chr,
  server_1L_chr = "dataverse.harvard.edu",
  key_1L_chr = NULL
)
}
\arguments{
\item{ids_chr}{Identities (a character vector)}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\value{
Dataverse datasets model summarys (a list)
}
\description{
get_dv_dss_mdl_smrys() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get dataverse datasets model summarys. Function argument ids_chr specifies the where to look for the required object. The function returns Dataverse datasets model summarys (a list).
}
\keyword{internal}
