% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_filtered_ttu_dss}
\alias{get_filtered_ttu_dss}
\title{Get filtered transfer to utility algorithm datasets}
\usage{
get_filtered_ttu_dss(
  ttu_dv_dss_tb = NULL,
  mdl_predrs_in_ds_chr = NULL,
  utility_type_chr = NULL,
  ttu_dv_nms_chr = "TTU",
  server_1L_chr = "dataverse.harvard.edu",
  key_1L_chr = NULL
)
}
\arguments{
\item{ttu_dv_dss_tb}{Transfer to utility algorithm dataverse datasets (a tibble), Default: NULL}

\item{mdl_predrs_in_ds_chr}{Model predictors in dataset (a character vector), Default: NULL}

\item{utility_type_chr}{Utility type (a character vector), Default: NULL}

\item{ttu_dv_nms_chr}{Transfer to utility algorithm dataverse names (a character vector), Default: 'TTU'}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\value{
Transfer to utility algorithm dataverse datasets (a tibble)
}
\description{
get_filtered_ttu_dss() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get filtered transfer to utility algorithm datasets. The function returns Transfer to utility algorithm dataverse datasets (a tibble).
}
\keyword{internal}
