% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_mdl_catalogue_refs}
\alias{get_mdl_catalogue_refs}
\title{Get model catalogue references}
\usage{
get_mdl_catalogue_refs(predictors_chr, ingredients_ls)
}
\arguments{
\item{predictors_chr}{Predictors (a character vector)}

\item{ingredients_ls}{Ingredients (a list)}
}
\value{
Catalogue references (a character vector)
}
\description{
get_mdl_catalogue_refs() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get model catalogue references. Function argument predictors_chr specifies the where to look for the required object. The function returns Catalogue references (a character vector).
}
\keyword{internal}
