% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_mdl_ctlg_url}
\alias{get_mdl_ctlg_url}
\title{Get model catalogue url}
\usage{
get_mdl_ctlg_url(
  mdls_lup,
  mdl_nm_1L_chr,
  server_1L_chr = "dataverse.harvard.edu",
  key_1L_chr = NULL
)
}
\arguments{
\item{mdls_lup}{Models (a lookup table)}

\item{mdl_nm_1L_chr}{Model name (a character vector of length one)}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\value{
ctlg_url (An object)
}
\description{
get_mdl_ctlg_url() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get model catalogue url. The function is called for its side effects and does not return a value.
}
\keyword{internal}
