% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_mdl_from_dv}
\alias{get_mdl_from_dv}
\title{Get mdl from dataverse}
\usage{
get_mdl_from_dv(
  mdl_nm_1L_chr,
  dv_ds_nm_1L_chr = "https://doi.org/10.7910/DVN/JC6PTV",
  server_1L_chr = "dataverse.harvard.edu",
  key_1L_chr = NULL
)
}
\arguments{
\item{mdl_nm_1L_chr}{Mdl name (a character vector of length one)}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'https://doi.org/10.7910/DVN/JC6PTV'}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\value{
NA ()
}
\description{
get_mdl_from_dv() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get mdl from dataverse. Function argument mdl_nm_1L_chr specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
