% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_mdl_metadata}
\alias{get_mdl_metadata}
\title{Get model metadata}
\usage{
get_mdl_metadata(
  mdls_lup,
  mdl_nm_1L_chr,
  server_1L_chr = "dataverse.harvard.edu",
  key_1L_chr = NULL
)
}
\arguments{
\item{mdls_lup}{Models (a lookup table)}

\item{mdl_nm_1L_chr}{Model name (a character vector of length one)}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\value{
Ingredients (a list)
}
\description{
get_mdl_metadata() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get model metadata. The function returns Ingredients (a list).
}
\keyword{internal}
