% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_mdl_smrys}
\alias{get_mdl_smrys}
\title{Get model summarys}
\usage{
get_mdl_smrys(ingredients_ls, mdl_nms_chr = NULL)
}
\arguments{
\item{ingredients_ls}{Ingredients (a list)}

\item{mdl_nms_chr}{Model names (a character vector), Default: NULL}
}
\value{
Models summary (a list)
}
\description{
get_mdl_smrys() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get model summarys. The function returns Models summary (a list).
}
\keyword{internal}
