% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_model}
\alias{get_model}
\title{Get model}
\usage{
get_model(
  mdls_lup,
  mdl_nm_1L_chr,
  make_from_tbl_1L_lgl = T,
  mdl_meta_data_ls = NULL,
  server_1L_chr = "dataverse.harvard.edu",
  key_1L_chr = NULL
)
}
\arguments{
\item{mdls_lup}{Models (a lookup table)}

\item{mdl_nm_1L_chr}{Model name (a character vector of length one)}

\item{make_from_tbl_1L_lgl}{Make from table (a logical vector of length one), Default: T}

\item{mdl_meta_data_ls}{Model meta data (a list), Default: NULL}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\value{
Model (a model)
}
\description{
get_model() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get model. The function returns Model (a model).
}
\keyword{internal}
