% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_predictors}
\alias{get_predictors}
\title{Get predictors}
\usage{
get_predictors(ingredients_ls)
}
\arguments{
\item{ingredients_ls}{Ingredients (a list)}
}
\value{
Predictors (a tibble)
}
\description{
get_predictors() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get predictors. Function argument ingredients_ls specifies the where to look for the required object. The function returns Predictors (a tibble).
}
\keyword{internal}
