% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_ttu_dv_dss}
\alias{get_ttu_dv_dss}
\title{Get transfer to utility algorithm dataverse datasets}
\usage{
get_ttu_dv_dss(
  ttu_dv_nms_chr = "TTU",
  server_1L_chr = "dataverse.harvard.edu",
  key_1L_chr = NULL
)
}
\arguments{
\item{ttu_dv_nms_chr}{Transfer to utility algorithm dataverse names (a character vector), Default: 'TTU'}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\value{
Transfer to utility algorithm dataverse datasets (a tibble)
}
\description{
get_ttu_dv_dss() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get transfer to utility algorithm dataverse datasets. Function argument ttu_dv_nms_chr specifies the where to look for the required object. The function returns Transfer to utility algorithm dataverse datasets (a tibble).
}
\keyword{internal}
