% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_ttu_dv_predrs}
\alias{get_ttu_dv_predrs}
\title{Get transfer to utility algorithm dataverse predictors}
\usage{
get_ttu_dv_predrs(
  ttu_dv_dss_tb = NULL,
  ttu_dv_nms_chr = "TTU",
  server_1L_chr = "dataverse.harvard.edu",
  key_1L_chr = NULL
)
}
\arguments{
\item{ttu_dv_dss_tb}{Transfer to utility algorithm dataverse datasets (a tibble), Default: NULL}

\item{ttu_dv_nms_chr}{Transfer to utility algorithm dataverse names (a character vector), Default: 'TTU'}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\value{
Predictors (a character vector)
}
\description{
get_ttu_dv_predrs() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get transfer to utility algorithm dataverse predictors. Function argument ttu_dv_dss_tb specifies the where to look for the required object. The function returns Predictors (a character vector).
}
\keyword{internal}
