% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_sofas.R
\name{is_firstbounce_sofas}
\alias{is_firstbounce_sofas}
\title{Is First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)}
\usage{
is_firstbounce_sofas(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\description{
Check whether an object is a valid instance of the First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\details{
First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
