% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_balanced_fake_ds}
\alias{make_balanced_fake_ds}
\title{Make balanced fake dataset}
\usage{
make_balanced_fake_ds(
  ds_tb,
  match_on_vars_chr,
  id_var_nm_1L_chr = "UID_chr",
  round_var_nm_1L_chr = "Timepoint_chr",
  timepoint_bl_val_1L_chr = "Baseline",
  cmprsn_var_nm_1L_chr = "study_arm_chr",
  cmprsn_groups_chr = c("Intervention", "Control")
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{match_on_vars_chr}{Match on vars (a character vector)}

\item{id_var_nm_1L_chr}{Id var name (a character vector of length one), Default: 'UID_chr'}

\item{round_var_nm_1L_chr}{Round var name (a character vector of length one), Default: 'Timepoint_chr'}

\item{timepoint_bl_val_1L_chr}{Timepoint bl value (a character vector of length one), Default: 'Baseline'}

\item{cmprsn_var_nm_1L_chr}{Cmprsn var name (a character vector of length one), Default: 'study_arm_chr'}

\item{cmprsn_groups_chr}{Cmprsn groups (a character vector), Default: c("Intervention", "Control")}
}
\value{
Dataset (a tibble)
}
\description{
make_balanced_fake_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make balanced fake dataset. The function returns Dataset (a tibble).
}
\keyword{internal}
