% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ce_smry}
\alias{make_ce_smry}
\title{Make ce summary}
\usage{
make_ce_smry(
  ds_tb,
  indices,
  change_types_chr = "dbl",
  benefits_pfx_1L_chr = "qalys_dbl",
  benefits_var_nm_1L_chr = "qalys",
  costs_pfx_1L_chr = "costs_dbl",
  costs_var_nm_1L_chr = "costs",
  change_sfx_1L_chr = "change",
  change_vars_chr = NA_character_,
  cmprsn_groups_chr = c("Intervention", "Control"),
  cmprsn_var_nm_1L_chr = "study_arm_chr",
  round_fup_1L_chr = "Follow-up"
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{indices}{PARAM_DESCRIPTION}

\item{change_types_chr}{Change types (a character vector), Default: 'dbl'}

\item{benefits_pfx_1L_chr}{Benefits prefix (a character vector of length one), Default: 'qalys_dbl'}

\item{benefits_var_nm_1L_chr}{Benefits var name (a character vector of length one), Default: 'qalys'}

\item{costs_pfx_1L_chr}{Costs prefix (a character vector of length one), Default: 'costs_dbl'}

\item{costs_var_nm_1L_chr}{Costs var name (a character vector of length one), Default: 'costs'}

\item{change_sfx_1L_chr}{Change suffix (a character vector of length one), Default: 'change'}

\item{change_vars_chr}{Change vars (a character vector), Default: 'NA'}

\item{cmprsn_groups_chr}{Cmprsn groups (a character vector), Default: c("Intervention", "Control")}

\item{cmprsn_var_nm_1L_chr}{Cmprsn var name (a character vector of length one), Default: 'study_arm_chr'}

\item{round_fup_1L_chr}{Round fup (a character vector of length one), Default: 'Follow-up'}
}
\value{
Summary (a double vector)
}
\description{
make_ce_smry() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make ce summary. The function returns Summary (a double vector).
}
\keyword{internal}
