% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_costs_vec_from_gamma_dist}
\alias{make_costs_vec_from_gamma_dist}
\title{Make costs vector from gamma dist}
\usage{
make_costs_vec_from_gamma_dist(n_int, costs_mean_dbl, costs_sd_dbl)
}
\arguments{
\item{n_int}{N (an integer vector)}

\item{costs_mean_dbl}{Costs mean (a double vector)}

\item{costs_sd_dbl}{Costs sd (a double vector)}
}
\value{
Costs (a double vector)
}
\description{
make_costs_vec_from_gamma_dist() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make costs vector from gamma dist. The function returns Costs (a double vector).
}
\keyword{internal}
