% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_costs_vec_from_gamma_dstr}
\alias{make_costs_vec_from_gamma_dstr}
\title{Make costs vector from gamma distribution}
\usage{
make_costs_vec_from_gamma_dstr(n_int, costs_mean_dbl, costs_sd_dbl)
}
\arguments{
\item{n_int}{N (an integer vector)}

\item{costs_mean_dbl}{Costs mean (a double vector)}

\item{costs_sd_dbl}{Costs standard deviation (a double vector)}
}
\value{
Costs (a double vector)
}
\description{
make_costs_vec_from_gamma_dstr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make costs vector from gamma distribution. The function returns Costs (a double vector).
}
\keyword{internal}
