% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fake_ds_one}
\alias{make_fake_ds_one}
\title{Make fake dataset one}
\usage{
make_fake_ds_one()
}
\value{
Fake data (a tibble)
}
\description{
make_fake_ds_one() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make fake dataset one. The function returns Fake data (a tibble).
}
\keyword{internal}
