% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fake_ds_two}
\alias{make_fake_ds_two}
\title{Make fake dataset two}
\usage{
make_fake_ds_two()
}
\value{
Matched dataset (a tibble)
}
\description{
make_fake_ds_two() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make fake dataset two. The function returns Matched dataset (a tibble).
}
\keyword{internal}
