% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_he_smry}
\alias{make_he_smry}
\title{Make he smry}
\usage{
make_he_smry(
  ds_tb,
  change_vars_chr = NA_character_,
  wtp_dbl = 50000,
  bootstrap_iters_1L_int = 1000,
  change_types_chr = "dbl",
  benefits_pfx_1L_chr = "qalys_dbl",
  benefits_var_nm_1L_chr = "qalys",
  costs_pfx_1L_chr = "costs_dbl",
  costs_var_nm_1L_chr = "costs",
  change_sfx_1L_chr = "change",
  cmprsn_groups_chr = c("Intervention", "Control"),
  cmprsn_var_nm_1L_chr = "study_arm_chr",
  round_fup_1L_chr = "Follow-up"
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{change_vars_chr}{Change vars (a character vector), Default: 'NA'}

\item{wtp_dbl}{Wtp (a double vector), Default: 50000}

\item{bootstrap_iters_1L_int}{Bootstrap iters (an integer vector of length one), Default: 1000}

\item{change_types_chr}{Change types (a character vector), Default: 'dbl'}

\item{benefits_pfx_1L_chr}{Benefits prefix (a character vector of length one), Default: 'qalys_dbl'}

\item{benefits_var_nm_1L_chr}{Benefits var name (a character vector of length one), Default: 'qalys'}

\item{costs_pfx_1L_chr}{Costs prefix (a character vector of length one), Default: 'costs_dbl'}

\item{costs_var_nm_1L_chr}{Costs var name (a character vector of length one), Default: 'costs'}

\item{change_sfx_1L_chr}{Change suffix (a character vector of length one), Default: 'change'}

\item{cmprsn_groups_chr}{Cmprsn groups (a character vector), Default: c("Intervention", "Control")}

\item{cmprsn_var_nm_1L_chr}{Cmprsn var name (a character vector of length one), Default: 'study_arm_chr'}

\item{round_fup_1L_chr}{Round fup (a character vector of length one), Default: 'Follow-up'}
}
\value{
He smry (a list)
}
\description{
make_he_smry() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make he smry. The function returns He smry (a list).
}
\keyword{internal}
