% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_hlth_ec_smry}
\alias{make_hlth_ec_smry}
\title{Make health economic summary}
\usage{
make_hlth_ec_smry(
  ds_tb,
  predn_ds_ls,
  wtp_dbl = 50000,
  bootstrap_iters_1L_int = 1000,
  benefits_pfx_1L_chr = "qalys_dbl",
  benefits_var_nm_1L_chr = "qalys",
  costs_var_nm_1L_chr = "costs",
  change_sfx_1L_chr = "change"
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{predn_ds_ls}{Prediction dataset (a list)}

\item{wtp_dbl}{Willingness to pay (a double vector), Default: 50000}

\item{bootstrap_iters_1L_int}{Bootstrap iterations (an integer vector of length one), Default: 1000}

\item{benefits_pfx_1L_chr}{Benefits prefix (a character vector of length one), Default: 'qalys_dbl'}

\item{benefits_var_nm_1L_chr}{Benefits variable name (a character vector of length one), Default: 'qalys'}

\item{costs_var_nm_1L_chr}{Costs variable name (a character vector of length one), Default: 'costs'}

\item{change_sfx_1L_chr}{Change suffix (a character vector of length one), Default: 'change'}
}
\value{
He summary (a list)
}
\description{
make_hlth_ec_smry() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make health economic summary. The function returns He summary (a list).
}
