% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_matched_ds}
\alias{make_matched_ds}
\title{Make matched dataset}
\usage{
make_matched_ds(sngl_grp_ds, cmprsn_smry_tb, ds_smry_ls)
}
\arguments{
\item{sngl_grp_ds}{PARAM_DESCRIPTION}

\item{cmprsn_smry_tb}{Cmprsn smry (a tibble)}

\item{ds_smry_ls}{Dataset smry (a list)}
}
\value{
Matched dataset (a tibble)
}
\description{
make_matched_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make matched dataset. The function returns Matched dataset (a tibble).
}
