% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_sofas.R
\name{make_pt_firstbounce_sofas}
\alias{make_pt_firstbounce_sofas}
\title{Make prototype First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)}
\usage{
make_pt_firstbounce_sofas()
}
\value{
A prototype for First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\description{
Create a new prototype for the First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\details{
First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
