% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_valid_predn_ds_ls}
\alias{make_valid_predn_ds_ls}
\title{Make valid prediction dataset}
\usage{
make_valid_predn_ds_ls(
  data_tb,
  id_var_nm_1L_chr = "UID",
  mdl_meta_data_ls = NULL,
  mdls_lup = NULL,
  mdl_nm_1L_chr = NULL,
  msrmnt_date_var_nm_1L_chr = NULL,
  predr_vars_nms_chr = NULL,
  round_var_nm_1L_chr,
  round_bl_val_1L_chr,
  utl_var_nm_1L_chr = "AQoL6D_HU",
  server_1L_chr = "dataverse.harvard.edu",
  key_1L_chr = NULL
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one), Default: 'UID'}

\item{mdl_meta_data_ls}{Model meta data (a list), Default: NULL}

\item{mdls_lup}{Models (a lookup table), Default: NULL}

\item{mdl_nm_1L_chr}{Model name (a character vector of length one), Default: NULL}

\item{msrmnt_date_var_nm_1L_chr}{Measurement date variable name (a character vector of length one), Default: NULL}

\item{predr_vars_nms_chr}{Predictor variables names (a character vector), Default: NULL}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one)}

\item{round_bl_val_1L_chr}{Round baseline value (a character vector of length one)}

\item{utl_var_nm_1L_chr}{Utility variable name (a character vector of length one), Default: 'AQoL6D_HU'}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\value{
Valid prediction dataset (a list)
}
\description{
make_valid_predn_ds_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make valid prediction dataset list. The function returns Valid prediction dataset (a list).
}
\keyword{internal}
