% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_mdls_lup.R
\docType{data}
\name{mdls_lup}
\alias{mdls_lup}
\title{Lookup table of prediction models}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24 rows and 4 columns.
}
\usage{
mdls_lup
}
\description{
A summary of the key descriptive features of the prediction models included in the youthu package.
}
\details{
A tibble

\describe{
\item{mdl_nms_chr}{Model names (a character vector)}
\item{predrs_ls}{Predictors (a list)}
\item{mdl_type_chr}{Model type (a character vector)}
\item{tfmn_chr}{Transformation (a character vector)}
}
}
\keyword{datasets}
