% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_mdls_smry_tb.R
\docType{data}
\name{mdls_smry_tb}
\alias{mdls_smry_tb}
\title{Summary coefficients for youthu transfer to utility models}
\format{
An object of class \code{data.frame} with 168 rows and 5 columns.
}
\usage{
mdls_smry_tb
}
\description{
A summary of the models included in the youthu package that can be used to predict adolescent AQoL6D. Note this summary is a placeholder as all models have been estimated from synthetic data.
}
\details{
A tibble

\describe{
\item{Model}{NO MATCH}
\item{Parameter}{NO MATCH}
\item{Estimate}{NO MATCH}
\item{SE}{NO MATCH}
\item{CI_95}{NO MATCH}
}
}
\keyword{datasets}
