% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_ds_for_cmprsn}
\alias{transform_ds_for_cmprsn}
\title{Transform dataset for comparison}
\usage{
transform_ds_for_cmprsn(
  ds_tb,
  cmprsn_var_nm_1L_chr,
  id_var_nm_1L_chr = "UID_chr",
  round_var_nm_1L_chr = "Timepoint_chr",
  cmprsn_groups_chr = c("Intervention", "Control")
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{cmprsn_var_nm_1L_chr}{Comparison variable name (a character vector of length one)}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one), Default: 'UID_chr'}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one), Default: 'Timepoint_chr'}

\item{cmprsn_groups_chr}{Comparison groups (a character vector), Default: c("Intervention", "Control")}
}
\value{
Dataset (a tibble)
}
\description{
transform_ds_for_cmprsn() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform dataset for comparison. The function returns Dataset (a tibble).
}
\keyword{internal}
