% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_ds_to_drop_msng}
\alias{transform_ds_to_drop_msng}
\title{Transform dataset to drop missing}
\usage{
transform_ds_to_drop_msng(ds_tb, predictors_chr, uid_var_nm_1L_chr = "UID_chr")
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{predictors_chr}{Predictors (a character vector)}

\item{uid_var_nm_1L_chr}{Unique identifier variable name (a character vector of length one), Default: 'UID_chr'}
}
\value{
Dataset (a tibble)
}
\description{
transform_ds_to_drop_msng() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform dataset to drop missing. The function returns Dataset (a tibble).
}
\keyword{internal}
