% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_ds_to_predn_ds}
\alias{transform_ds_to_predn_ds}
\title{Transform dataset to prediction dataset}
\usage{
transform_ds_to_predn_ds(
  data_tb,
  predr_vars_nms_chr,
  tfmn_1L_chr,
  dep_var_nm_1L_chr = "aqol6d_total_w",
  id_var_nm_1L_chr = "fkClientID",
  round_var_nm_1L_chr = "round",
  round_bl_val_1L_chr = "Baseline",
  predictors_lup = NULL
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{predr_vars_nms_chr}{Predr vars names (a character vector)}

\item{tfmn_1L_chr}{Transformation (a character vector of length one)}

\item{dep_var_nm_1L_chr}{Dep var name (a character vector of length one), Default: 'aqol6d_total_w'}

\item{id_var_nm_1L_chr}{Id var name (a character vector of length one), Default: 'fkClientID'}

\item{round_var_nm_1L_chr}{Round var name (a character vector of length one), Default: 'round'}

\item{round_bl_val_1L_chr}{Round bl value (a character vector of length one), Default: 'Baseline'}

\item{predictors_lup}{Predictors (a lookup table), Default: NULL}
}
\value{
Data (a tibble)
}
\description{
transform_ds_to_predn_ds() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform dataset to prediction dataset. Function argument data_tb specifies the object to be updated. Argument predr_vars_nms_chr provides the object to be updated. The function returns Data (a tibble).
}
\keyword{internal}
