% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_col_with_diff}
\alias{update_col_with_diff}
\title{Update column with difference}
\usage{
update_col_with_diff(
  ds_tb,
  var_nm_1L_chr,
  fn,
  abs_mean_diff_1L_dbl,
  diff_sd_1L_dbl,
  multiplier_1L_dbl,
  min_1L_dbl,
  max_1L_dbl,
  integer_1L_lgl
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{var_nm_1L_chr}{Variable name (a character vector of length one)}

\item{fn}{Function (a function)}

\item{abs_mean_diff_1L_dbl}{Absolute mean difference (a double vector of length one)}

\item{diff_sd_1L_dbl}{Difference standard deviation (a double vector of length one)}

\item{multiplier_1L_dbl}{Multiplier (a double vector of length one)}

\item{min_1L_dbl}{Minimum (a double vector of length one)}

\item{max_1L_dbl}{Maximum (a double vector of length one)}

\item{integer_1L_lgl}{Integer (a logical vector of length one)}
}
\value{
New (a dataset)
}
\description{
update_col_with_diff() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update column with difference. The function is called for its side effects and does not return a value.
}
\keyword{internal}
