% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_multpl_cols_with_diffs}
\alias{update_multpl_cols_with_diffs}
\title{Update multpl columns with diffs}
\usage{
update_multpl_cols_with_diffs(
  ds_tb,
  var_nms_chr,
  fns_ls,
  abs_mean_diff_dbl,
  diff_sd_dbl,
  multiplier_dbl,
  min_dbl,
  max_dbl,
  integer_lgl
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{var_nms_chr}{Variable names (a character vector)}

\item{fns_ls}{Functions (a list)}

\item{abs_mean_diff_dbl}{Abs mean diff (a double vector)}

\item{diff_sd_dbl}{Diff sd (a double vector)}

\item{multiplier_dbl}{Multiplier (a double vector)}

\item{min_dbl}{Minimum (a double vector)}

\item{max_dbl}{Maximum (a double vector)}

\item{integer_lgl}{Integer (a logical vector)}
}
\value{
Updated dataset (a tibble)
}
\description{
update_multpl_cols_with_diffs() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update multpl columns with diffs. Function argument ds_tb specifies the object to be updated. Argument var_nms_chr provides the object to be updated. The function returns Updated dataset (a tibble).
}
\keyword{internal}
