% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_firstbounce_sofas.R
\name{validate_firstbounce_sofas}
\alias{validate_firstbounce_sofas}
\title{Validate First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)}
\usage{
validate_firstbounce_sofas(x)
}
\arguments{
\item{x}{An unvalidated instance of the First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)}
}
\value{
A prototpe for First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\description{
Validate an instance of the First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
\details{
First Bounce S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)
}
