% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_aqol6d_predn_to_ds}
\alias{add_aqol6d_predn_to_ds}
\title{Add aqol6d predn to dataset}
\usage{
add_aqol6d_predn_to_ds(
  data_tb,
  model_mdl,
  tfmn_1L_chr,
  predr_vars_nms_chr = NULL,
  utl_var_nm_1L_chr = NULL,
  id_var_nm_1L_chr = "fkClientID",
  round_var_nm_1L_chr = "round",
  round_bl_val_1L_chr = "Baseline",
  utl_cls_fn = firstbounce_aqol6d_adol,
  predictors_lup = NULL
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{model_mdl}{PARAM_DESCRIPTION}

\item{tfmn_1L_chr}{Tfmn (a character vector of length one)}

\item{predr_vars_nms_chr}{Predr vars names (a character vector), Default: NULL}

\item{utl_var_nm_1L_chr}{Utl var name (a character vector of length one), Default: NULL}

\item{id_var_nm_1L_chr}{Id var name (a character vector of length one), Default: 'fkClientID'}

\item{round_var_nm_1L_chr}{Round var name (a character vector of length one), Default: 'round'}

\item{round_bl_val_1L_chr}{Round bl value (a character vector of length one), Default: 'Baseline'}

\item{utl_cls_fn}{Utl class (a function), Default: firstbounce_aqol6d_adol}

\item{predictors_lup}{Predictors (a lookup table), Default: NULL}
}
\value{
Updated (a tibble)
}
\description{
add_aqol6d_predn_to_ds() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add aqol6d predn to dataset. Function argument data_tb specifies the object to be updated. The function returns Updated (a tibble).
}
