% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_aqol6d_predn_to_ds}
\alias{add_aqol6d_predn_to_ds}
\title{Add Assessment of Quality of Life Six Dimension prediction to dataset}
\usage{
add_aqol6d_predn_to_ds(
  data_tb,
  model_mdl,
  tfmn_1L_chr,
  predr_vars_nms_chr = NULL,
  utl_var_nm_1L_chr = NULL,
  id_var_nm_1L_chr = "fkClientID",
  round_var_nm_1L_chr = "round",
  round_bl_val_1L_chr = "Baseline",
  utl_cls_fn = youthvars::youthvars_aqol6d_adol,
  predictors_lup = NULL
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{model_mdl}{Model (a model)}

\item{tfmn_1L_chr}{Transformation (a character vector of length one)}

\item{predr_vars_nms_chr}{Predictor variables names (a character vector), Default: NULL}

\item{utl_var_nm_1L_chr}{Utility variable name (a character vector of length one), Default: NULL}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one), Default: 'fkClientID'}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one), Default: 'round'}

\item{round_bl_val_1L_chr}{Round baseline value (a character vector of length one), Default: 'Baseline'}

\item{utl_cls_fn}{Utility class (a function), Default: youthvars::youthvars_aqol6d_adol}

\item{predictors_lup}{Predictors (a lookup table), Default: NULL}
}
\value{
Updated (a tibble)
}
\description{
add_aqol6d_predn_to_ds() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add assessment of quality of life six dimension prediction to dataset. Function argument data_tb specifies the object to be updated. The function returns Updated (a tibble).
}
