% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_dv_mdl_smrys}
\alias{get_dv_mdl_smrys}
\title{Get dataverse model summarys}
\usage{
get_dv_mdl_smrys(mdls_lup, mdl_nms_chr = NULL)
}
\arguments{
\item{mdls_lup}{Models (a lookup table)}

\item{mdl_nms_chr}{Model names (a character vector), Default: NULL}
}
\value{
Dataverse model (summarys)
}
\description{
get_dv_mdl_smrys() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get dataverse model summarys. The function is called for its side effects and does not return a value.
}
\keyword{internal}
