% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_filtered_ttu_dss}
\alias{get_filtered_ttu_dss}
\title{Get filtered ttu datasets}
\usage{
get_filtered_ttu_dss(
  ttu_dv_dss_tb = NULL,
  mdl_predrs_in_ds_chr = NULL,
  utility_type_chr = NULL,
  ttu_dv_nm_1L_chr = "firstbounce",
  server_1L_chr = "dataverse.harvard.edu",
  key_1L_chr = NULL
)
}
\arguments{
\item{ttu_dv_dss_tb}{Ttu dataverse datasets (a tibble), Default: NULL}

\item{mdl_predrs_in_ds_chr}{Model predictors in dataset (a character vector), Default: NULL}

\item{utility_type_chr}{Utility type (a character vector), Default: NULL}

\item{ttu_dv_nm_1L_chr}{Ttu dataverse name (a character vector of length one), Default: 'firstbounce'}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\value{
Ttu dataverse datasets (a tibble)
}
\description{
get_filtered_ttu_dss() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get filtered ttu datasets. Function argument ttu_dv_dss_tb specifies the where to look for the required object. The function returns Ttu dataverse datasets (a tibble).
}
\keyword{internal}
