% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_mdl_ctlg_url}
\alias{get_mdl_ctlg_url}
\title{Get model catalogue url}
\usage{
get_mdl_ctlg_url(
  mdls_lup,
  mdl_nm_1L_chr,
  server_1L_chr = "dataverse.harvard.edu",
  key_1L_chr = NULL
)
}
\arguments{
\item{mdls_lup}{Models (a lookup table)}

\item{mdl_nm_1L_chr}{Model name (a character vector of length one)}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\value{
NA ()
}
\description{
get_mdl_ctlg_url() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get model catalogue url. Function argument mdls_lup specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
