% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_mdl_ds_url}
\alias{get_mdl_ds_url}
\title{Get model dataset url}
\usage{
get_mdl_ds_url(mdls_lup, mdl_nm_1L_chr)
}
\arguments{
\item{mdls_lup}{Models (a lookup table)}

\item{mdl_nm_1L_chr}{Model name (a character vector of length one)}
}
\value{
mdl_ds_url (An object)
}
\description{
get_mdl_ds_url() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get model dataset url. The function is called for its side effects and does not return a value.
}
\keyword{internal}
