% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_mdl_from_dv}
\alias{get_mdl_from_dv}
\title{Get model from dataverse}
\usage{
get_mdl_from_dv(
  mdl_nm_1L_chr,
  dv_ds_nm_1L_chr = "https://doi.org/10.7910/DVN/JC6PTV",
  server_1L_chr = "dataverse.harvard.edu",
  key_1L_chr = NULL
)
}
\arguments{
\item{mdl_nm_1L_chr}{Model name (a character vector of length one)}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'https://doi.org/10.7910/DVN/JC6PTV'}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\value{
Model (a model)
}
\description{
get_mdl_from_dv() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get model from dataverse. The function returns Model (a model).
}
