% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_mdls_using_predrs}
\alias{get_mdls_using_predrs}
\title{Get models using predictors}
\usage{
get_mdls_using_predrs(mdl_predrs_in_ds_chr, mdls_lup = NULL)
}
\arguments{
\item{mdl_predrs_in_ds_chr}{Model predictors in dataset (a character vector)}

\item{mdls_lup}{Models (a lookup table), Default: NULL}
}
\value{
Filtered models (a lookup table)
}
\description{
get_mdls_using_predrs() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get models using predictors. Function argument mdl_predrs_in_ds_chr specifies the where to look for the required object. The function returns Filtered models (a lookup table).
}
