% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_predictors_lup}
\alias{get_predictors_lup}
\title{Get predictors}
\usage{
get_predictors_lup(
  mdl_meta_data_ls = NULL,
  mdls_lup = NULL,
  mdl_nm_1L_chr = NULL,
  outp_is_abbrvs_tb = F,
  server_1L_chr = "dataverse.harvard.edu",
  key_1L_chr = NULL
)
}
\arguments{
\item{mdl_meta_data_ls}{Model meta data (a list), Default: NULL}

\item{mdls_lup}{Models (a lookup table), Default: NULL}

\item{mdl_nm_1L_chr}{Model name (a character vector of length one), Default: NULL}

\item{outp_is_abbrvs_tb}{Output is abbrvs (a tibble), Default: F}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\value{
Predictors (a tibble)
}
\description{
get_predictors_lup() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get predictors lookup table. Function argument mdl_meta_data_ls specifies the where to look for the required object. The function returns Predictors (a tibble).
}
\keyword{internal}
