% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_sngl_grp_ds}
\alias{make_sngl_grp_ds}
\title{Make sngl grp dataset}
\usage{
make_sngl_grp_ds(seed_ds_tb = NULL, ds_smry_ls)
}
\arguments{
\item{seed_ds_tb}{Seed dataset (a tibble), Default: NULL}

\item{ds_smry_ls}{Dataset summary (a list)}
}
\value{
Sngl grp dataset (a tibble)
}
\description{
make_sngl_grp_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make sngl grp dataset. The function returns Sngl grp dataset (a tibble).
}
