% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_youthu.R
\docType{package}
\name{youthu-package}
\alias{youthu}
\alias{youthu-package}
\title{youthu: Youth Outcomes to Health Utility}
\description{
Tools for mapping measures routinely collected in youth
mental health services to AQOL 6D Health Utility. Part of the First
Bounce model of primary youth mental health services.  This
development version of the youthu package has been made available as
part of the process of testing and documenting the package. The tools
contained in this development release are designed for use in
conjunction with model objects stored in data repositories. The real
model objects will be publicly released once the associated scientific
manuscript is published. In the mean time, we have included links to
placeholder model objects derived from synthetic data.  For this
reason, this release is for demonstration purposes only and this
package should not yet be used in analyses deigned to inform policy
decisions. If you have any questions, please contact the authors
(matthew.hamilton@orygen.org.au).  The documentation for this package
has been automatically generated by the ready4fun package and is
therefore quite rudimentary. Human authored documentation will follow
in 2021.
}
\details{
To learn more about youthu, start with the vignettes:
\code{browseVignettes(package = "youthu")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/youthu/}
  \item \url{https://github.com/ready4-dev/youthu}
  \item \url{https://www.ready4-dev.com/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Authors:
\itemize{
  \item Caroline Gao \email{caroline.gao@orygen.org.au} (\href{https://orcid.org/0000-0002-0987-2759}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item Headspace [funder]
  \item National Health and Medical Research Council [funder]
}

}
